+++ Tersan.SketchManagement/Infrastructure/Persistence/ViewModels/Building/BuildingSummaryViewModel.cs
@@ -6,8 +6,8 @@
 
         public string? Name { get; set; }
 
-        public int X { get; set; }
+        public decimal X { get; set; }
 
-        public int Y { get; set; }
+        public decimal Y { get; set; }
     }
 }
