+++ Tersan.SketchManagement/Infrastructure/Persistence/ViewModels/Ship/InputShipViewModel.cs
@@ -15,9 +15,9 @@
 
         public string? ShipStatus { get; set; }
 
-        public int X { get; set; }
+        public decimal X { get; set; }
 
-        public int Y { get; set; }
+        public decimal Y { get; set; }
 
     }
 }
