+++ Tersan.SketchManagement/Infrastructure/Persistence/ViewModels/Ship/OutputShipViewModel.cs
@@ -5,8 +5,8 @@
         public int ID { get; set; }
         public string? Name { get; set; }
         public int ShipStatusID { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public decimal X { get; set; }
+        public decimal Y { get; set; }
         public string? StatusType { get; set; }
     }
 }
